requirements.txt
flask==3.1.0
google-cloud-vision==3.9.0
google-cloud-texttospeech==2.21.1
google-cloud-storage==2.19.0
google-cloud-org-policy==1.11.0
EbookLib==0.18
beautifulsoup4==4.12.3
lxml==5.3.0
Pillow==11.1.0
python-dotenv==1.0.1
gunicorn==23.0.0
Werkzeug==3.1.3
